Remove forced white background on gamers list
On the dark theme, the list of gamers ends up with a white background.

Removing the CSS background-color element results in them taking on the
background color from the theme.

This has no effect on the light/default theme since it is setting the
background to the default background anyway.

The hover highlight color is a bit jaring on the dark theme, but that
will take more experimentation to find a good color and should go in
themeDark.less not here.