Remove forced white background on profile pages
On the dark theme, user's avatar ends up with a white background. If the
image has transparency this will be misleading as that is not how it
will display elsewhere.

Removing the CSS background-color element results in them taking on the
background color from the theme.

This has no effect on the light/default theme since it is setting the
background to the default background anyway.